add monkey.ts
